fix(qa): split suite runtime agent process below line cap
