+++ extensions/qa-lab/src/suite-runtime-agent-process.integration.test.ts
@@ -3,7 +3,7 @@ import { mkdir, mkdtemp, rm, writeFile } from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
 import { afterEach, describe, expect, it } from "vitest";
-import { runQaCli } from "./suite-runtime-agent-process.js";
+import { runQaCli } from "./qa-cli-process.js";
 
 const cleanups: Array<() => Promise<void>> = [];
